Year 2 Block D/Azure_edo/airflow/requirements.txt
apache-airflow==2.9.1
azure-identity==1.17.0
azure-storage-blob==12.19.0
azure-ai-ml
azure-core
pandas
scikit-learn
